% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonDB.R
\name{getAttr}
\alias{getAttr}
\title{genAttr}
\usage{
getAttr(database, mart, filters, value, cond)
}
\arguments{
\item{database}{RSQLite object imported previously from an SQLite database file to R}

\item{mart}{Dataset table into the database, use the syntax "dataset"}

\item{filters}{A character vector with filters with the restrictions of the query restriction of the query}

\item{value}{A character vector with the corresponding values for filters}

\item{cond}{Conditional when more than one filter is used "and", "or","not
#' @examples ## import the SQLite database
## path_sqlite3 <- file.choose()
regulon <- DBI::dbConnect(RSQLite::SQLite(), path_sqlite3)
listDatasets(regulon)
listAttributes(regulon, "TF")
getAttr(
database = regulon,
mart = "GENE",
filters = c("operon_id", "operon_name"),
value = c("ECK125235983","cmoM-mukFEB" ),
cond = "and")
GetRegulatedGenesByTF(regulon, "Ara")
 dbDisconnect(regulon)}
}
\description{
Allows users to filter and retrieve data from a particular mart from SQLite database.
}
