% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonDB.R
\name{getGeneRegulation}
\alias{getGeneRegulation}
\title{getGeneRegulation}
\usage{
getGeneRegulation(database, genes, format, output.type)
}
\arguments{
\item{database}{RSQLite object imported previously from an SQLite database file to R}

\item{genes}{genes of interest}

\item{format}{onerow, multirow, table}

\item{output.type}{output object
#' @examples ## import the SQLite database
## path_sqlite3 <- file.choose()
regulon <- DBI::dbConnect(RSQLite::SQLite(), path_sqlite3)
listDatasets(regulon)
listAttributes(regulon, "TF")
getAttr(
database = regulon,
mart = "GENE",
filters = c("operon_id", "operon_name"),
value = c("ECK125235983","cmoM-mukFEB" ),
cond = "and")
GetRegulatedGenesByTF(regulon, "Ara")
 dbDisconnect(regulon)}

\item{mart}{Dataset table into the database, use the syntax "dataset"}
}
\description{
provides gene regulation
}
