% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonDB.R
\name{GetRegulatedGenesByTF}
\alias{GetRegulatedGenesByTF}
\title{##GenRegulatedGenesByTF}
\usage{
GetRegulatedGenesByTF(database, tf)
}
\arguments{
\item{database}{RSQLite object imported previously from an SQLite database file to R}

\item{mart}{Dataset table into the database, use the syntax "dataset"}
}
\description{
##GenRegulatedGenesByTF
}
