% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonDB.R
\name{getAttr}
\alias{getAttr}
\title{##genAttr}
\usage{
getAttr(database, mart, filters, value, cond)
}
\arguments{
\item{database}{RSQLite object imported previously from an SQLite database file to R}

\item{mart}{Dataset table into the database, use the syntax "dataset"}

\item{filters}{A character vector with filters with the restrictions of the query restriction of the query}

\item{value}{A character vector with the corresponding values for filters}

\item{cond}{Conditional when more than one filter is used "and", "or","not}
}
\description{
##genAttr
}
