% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonDB.R
\name{getGeneRegulation}
\alias{getGeneRegulation}
\title{##getGeneRegulation}
\usage{
getGeneRegulation(database, genes, format, output.type)
}
\arguments{
\item{database}{RSQLite object imported previously from an SQLite database file to R}

\item{genes}{genes of interest}

\item{format}{onerow, multirow, table}

\item{output.type}{output object}

\item{mart}{Dataset table into the database, use the syntax "dataset"}
}
\description{
##getGeneRegulation
}
